requirements.txt
Django
requests
beautifulsoup4
python-slugify
Pillow
pyyaml